requirements.txt
websocket_client